+++ Tests/RxGRDBTests/DatabaseWriterWriteTests.swift
@@ -99,7 +99,7 @@ class DatabaseWriterWriteTests : XCTestCase {
         }
         
         try Test(test)
-            .run { DatabaseQueue() }
+            .run { try DatabaseQueue() }
             .runAtTemporaryDatabasePath { try DatabaseQueue(path: $0) }
             .runAtTemporaryDatabasePath { try DatabasePool(path: $0) }
     }
@@ -279,7 +279,7 @@ class DatabaseWriterWriteTests : XCTestCase {
         }
         
         try Test(test)
-            .run { DatabaseQueue() }
+            .run { try DatabaseQueue() }
             .runAtTemporaryDatabasePath { try DatabaseQueue(path: $0) }
             .runAtTemporaryDatabasePath { try DatabasePool(path: $0) }
     }
@@ -301,7 +301,7 @@ class DatabaseWriterWriteTests : XCTestCase {
         }
         
         try Test(test)
-            .run { DatabaseQueue() }
+            .run { try DatabaseQueue() }
             .runAtTemporaryDatabasePath { try DatabaseQueue(path: $0) }
             .runAtTemporaryDatabasePath { try DatabasePool(path: $0) }
     }
@@ -353,7 +353,7 @@ class DatabaseWriterWriteTests : XCTestCase {
         }
         
         try Test(test)
-            .run { DatabaseQueue() }
+            .run { try DatabaseQueue() }
             .runAtTemporaryDatabasePath { try DatabaseQueue(path: $0) }
             .runAtTemporaryDatabasePath { try DatabasePool(path: $0) }
     }